requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==1.11.16
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
idna==2.7
Markdown==3.0.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
django-model-utils==2.0.3
Pillow
matplotlib